feat/#39: size-write 뷰에서 이동했을 때 Layout에서 뷰 크기 조절
