+++ src/hooks/queries/useRemoveLocalStorage.ts
@@ -1,0 +1,7 @@
+export const useRemoveLocalStorage = () => {
+  const removeLocalStorageItem = (...args: unknown[]) => {
+    args.forEach((key) => localStorage.removeItem(`${key}`));
+  };
+
+  return { removeLocalStorageItem };
+};
