+++ src/hooks/ui/useGoBackToHistory.ts
@@ -2,21 +2,12 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 
 import { currentViewState, historyState } from '../../states/atom';
 
-const useGoBackToHistory = () => {
+export const useGoBackToHistory = () => {
   const [, setCurrentView] = useRecoilState(currentViewState);
   const history = useRecoilValue(historyState);
 
-  const goBacktoHistory = () => {
-    document.body.style.width = '38rem';
-    document.body.style.height = '37.5rem';
-    const container = document.getElementById('app-container') as HTMLElement;
-    if (container) {
-      container.style.width = '38rem';
-      container.style.height = '37.5rem';
-    }
+  const goBackToHistory = () => {
     history && setCurrentView(history);
   };
-  return goBacktoHistory;
+  return { goBackToHistory };
 };
-
-export default useGoBackToHistory;
